pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "mcp-ingest"
version = "0.1.0"
description = "SDK + CLI to describe, validate, and register MCP servers into MatrixHub"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "MatrixHub Contributors" }]
keywords = ["mcp", "model-context-protocol", "matrixhub", "ingest", "agents"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "httpx>=0.24,<0.29",
    "pydantic>=2,<3",
    "anyio>=3.6,<5",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.8",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "build>=1.0",
    "twine>=4.0",
    "types-PyYAML",
]
harvester = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "SQLAlchemy>=2.0",
]
# --- New: docs extras for MkDocs site ---------------------------------------
# Matches Makefile's docs-setup target; lets you `pip install -e ".[dev,harvester,docs]"`
docs = [
    "mkdocs>=1.5,<2",
    "mkdocs-material>=9.5,<10",
    "mkdocs-mermaid2-plugin>=1.1,<2",
    # Material uses pymdown-extensions; include explicitly for local builds
    "pymdown-extensions>=10,<12",
]

[project.scripts]
mcp-ingest = "mcp_ingest.cli:main"

# Ruff configuration (top-level applies to formatter and linter)
[tool.ruff]
line-length = 100

# Linter-specific configuration
[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
# Add all error codes you want to temporarily skip to this list
ignore = [
    "E501",  # Line too long
    "E402",  # Module level import not at top of file
    "UP022", # Prefer `capture_output` for subprocess
    "E731",  # Do not assign a `lambda` expression
    "N805",  # False positive: first argument of a method should be `self`
]
[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"